+++ docs/changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+--
+
+## 0.14.0
 * Bump Flank version to 21.01.1
 * Add support for test-targets-for-shard [Fixes #205](https://github.com/runningcode/fladle/issues/205)
 * Add support for new flank options [PR#211](https://github.com/runningcode/fladle/pull/211) Thanks [pawelpasterz](https://github.com/pawelpasterz)